Create Module for default custom value types
